generic: Implement busy delay for DelayMs<u8>
Implement DelayMs<u8> as well, for compatibility with drivers that
require u8 instead of u16.

+++ boards/arduino-uno/examples/uno-i2cdetect.rs
@@ -33,6 +33,6 @@ fn main() -> ! {
     i2c.i2cdetect(&mut serial, arduino_uno::hal::i2c::Direction::Read).void_unwrap();
 
     loop {
-        delay.delay_ms(1000);
+        delay.delay_ms(1000 as u16);
     }
 }
