+++ boards/bigavr6/examples/bigavr6-i2cdetect.rs
@@ -32,6 +32,6 @@ fn main() -> ! {
     i2c.i2cdetect(&mut serial, bigavr6::hal::i2c::Direction::Read).void_unwrap();
 
     loop {
-        delay.delay_ms(1000);
+        delay.delay_ms(1000 as u16);
     }
 }
